+++ indico/modules/rb/util.py
@@ -275,7 +275,7 @@ def get_booking_params_for_event(event):
         'text': event.room.name if event.room else None,
     }
     all_times = {day: (_find_first_entry_start_dt(event, day), _find_latest_entry_end_dt(event, day))
-                 for day in event.iter_days()}
+                 for day in event.iter_days(tzinfo=event.tzinfo)}
     # if the timetable is empty on a given day, use (start_dt, end_dt) of the event
     all_times = [((day, (event.start_dt_local, event.end_dt_local)) if times[0] is None else (day, times))
                  for day, times in all_times.viewitems()]